comments resolving
