+++ src/main/java/fr/inria/corese/core/next/query/impl/sparql/bridge/CoreseAstQueryBuilder.java
@@ -1,11 +1,14 @@
 package fr.inria.corese.core.next.query.impl.sparql.bridge;
 
+import fr.inria.corese.core.next.query.impl.sparql.ast.ConstraintAst;
+import fr.inria.corese.core.next.query.impl.sparql.ast.FilterAst;
 import fr.inria.corese.core.next.query.impl.sparql.ast.TermAst;
 import fr.inria.corese.core.next.query.kgram.api.core.Filter;
 
+import java.util.Objects;
+
 /**
  * Builds KGRAM {@code Exp} / {@code Query} structures from Corese-next query AST nodes.
- *
  */
 public final class CoreseAstQueryBuilder {
 
@@ -13,13 +16,32 @@ public final class CoreseAstQueryBuilder {
     }
 
     /**
-     * Converts a filter term (boolean expression) from the next parser into a
-     * {@link fr.inria.corese.core.sparql.triple.parser.Expression}, then exposes it as a
-     * {@link Filter} with {@link Filter#coreseNextSource()} set.
-     *
-     * <p>For static-style calls without an instance, use {@link SparqlAstToExpression#toNextFilter(TermAst)}.
+     * Converts a SPARQL {@code FILTER} clause by converting {@link FilterAst#operator()} the same way as
+     * {@link #toNextFilter(TermAst)}.
+     */
+    public Filter toNextFilter(FilterAst filterClause) {
+        Objects.requireNonNull(filterClause, "filterClause");
+        return toNextFilter(filterClause.operator());
+    }
+
+    /**
+     * Converts a filter expression carried as {@link TermAst}: must be a {@link ConstraintAst}.
      */
     public Filter toNextFilter(TermAst filterExpression) {
+        Objects.requireNonNull(filterExpression, "filterExpression");
+        if (!(filterExpression instanceof ConstraintAst constraint)) {
+            throw new IllegalArgumentException(
+                    "FILTER expects a ConstraintAst, got: " + filterExpression.getClass().getName());
+        }
+        return toNextFilter(constraint);
+    }
+
+    /**
+     * Converts a constraint tree (boolean filter expression) into a KGRAM {@link Filter}.
+     */
+    public Filter toNextFilter(ConstraintAst filterExpression) {
+        Objects.requireNonNull(filterExpression, "filterExpression");
         return SparqlAstToExpression.toNextFilter(filterExpression);
     }
+
 }
