+++ src/main/java/fr/inria/corese/core/next/util/StringUtils.java
@@ -1,5 +1,7 @@
 package fr.inria.corese.core.next.util;
 
+import fr.inria.corese.core.sparql.triple.parser.Processor;
+
 public class StringUtils {
 
     public static String trimChevronIRIs(String uri) {
@@ -93,5 +95,29 @@ public class StringUtils {
             }
         }
         return sb.toString();
+    }
+
+    /**
+     * Strips angle brackets if present, then returns the local part after {@code #}, {@code /}, or {@code :},
+     * lowercased for {@link Processor} lookup.
+     */
+    public static String localNameFromIriToken(String raw) {
+        String t = raw.trim();
+        if (t.length() >= 2 && t.charAt(0) == '<' && t.charAt(t.length() - 1) == '>') {
+            t = t.substring(1, t.length() - 1);
+        }
+        int hash = t.lastIndexOf('#');
+        if (hash >= 0 && hash < t.length() - 1) {
+            return t.substring(hash + 1).toLowerCase();
+        }
+        int slash = t.lastIndexOf('/');
+        if (slash >= 0 && slash < t.length() - 1) {
+            return t.substring(slash + 1).toLowerCase();
+        }
+        int colon = t.lastIndexOf(':');
+        if (colon >= 0 && colon < t.length() - 1) {
+            return t.substring(colon + 1).toLowerCase();
+        }
+        return t.toLowerCase();
     }
 }
